Create README.md
This is the Capstone Project for the DataTrained Platform. I have done the project in three parts. The was collected in Excel files. The initial analysis was done in excel. The initial description was written in Word Doc. The graphs generate in Excel were pasted in the same Word Doc. The other phases were done in Jupyter notebook and are imported in this repository.